Allows for comments when flipping to a grid
This resolves a bug where having comments in the variables or other config files would cause those lines to break.  This maintains the comments as a list at the end of the section.
This also resolves the sort order for the labels so that the longest ones are matched first when other labels contain a subset of previously defined ones.